fix: 修复自定义组件内删除时会删除最后一个 textNode，导致无法再次编辑

+++ lib/src/service/internal_key_event_handlers/backspace_handler.dart
@@ -19,6 +19,16 @@ ShortcutEventHandler backspaceEventHandler = (editorState, event) {
   final transaction = editorState.transaction;
   List<int>? cancelNumberListPath;
 
+  final parentNode = nodes.last.parent;
+  final currentTextNode = nodes.last;
+
+  if (currentTextNode is TextNode &&
+      currentTextNode.toPlainText() == "" &&
+      parentNode?.children.length == 1 &&
+      parentNode?.type != "editor") {
+    return KeyEventResult.handled;
+  }
+
   if (nonTextNodes.isNotEmpty) {
     transaction.deleteNodes(nonTextNodes);
   }
